requirements.txt
pandas==2.1.4
numpy>=1.26.0
prophet==1.1.5
requests==2.31.0
python-dotenv==1.0.0
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
